fix routing/activeClass issue
missing activeClassName quirk was fixed by nesting the analytics-admin-area and analytics-index views in a larger analytics view
now on both of these views the activeClassName persists. The next to-solve is making the div classes that wrap these components include the word
analytics and be unique. then we can close out this bug